Update verbose listing to nest aliases
This fixes a bug in the verbose listing: startswith is case-sensitive
and aliases begin with "Alg.Alias." not "Alg.alias" as presently
searched for.

Additionally, we update the verbose listing to nest aliases under the
respective "core" implementation heading. Both the implementation and
alias listings have been sorted, to be stable across multiple runs.

The new output of a verbose listing looks roughly like so:

    AlgorithmParameters : IvAlgorithmParameters
    AlgorithmParameters : RC2AlgorithmParameters
    AlgorithmParameters : RSAPSSAlgorithmParameters
    Cipher : AES
    Cipher : DES
    Cipher : DESede
        Alias: Cipher.DES3
    Cipher : RC2
    Cipher : RC4
    Cipher : RSA
    KeyFactory : DSA
    KeyFactory : EC
    KeyFactory : RSA
    KeyGenerator : AES
    KeyGenerator : DES
    KeyGenerator : DESede
        Alias: KeyGenerator.DES3

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>